requirements.txt
flask==3.0.2
flask-jwt-extended==4.6.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
flask-swagger-ui==4.11.1
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.4.3 